code: Visualize considering time axis.
In previous I just plotted the spots in volin-plot with out considering the time.

But since our classification needs to be studied using the pattern of current on time, I have changed my code to plot data in current-time graph.

This code only plot's data that is sensored in "Heptaol" scent.

Future Task : Plot all the data from 8 scents.