code: Making data in to same shape.
In order to put data in to DNN, the shape of array should be same.

In this data most of the trial has data length 1700~4000, But the some of the trials has too large, small amount of data!

So I have ignored those data(e.g first(0), last trials and 1st trial in 1-heptanol.